requirements.txt
Flask==3.0.0
wikipedia-api==0.6.0
nltk==3.8.1
aiohttp==3.9.1
python-dotenv==1.0.0
